Separate Login as a different component from Home
Add Login.js
