+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/home/apoorv/todo-frontend/src/App.js":"1","/home/apoorv/todo-frontend/src/components/Home.js":"2","/home/apoorv/todo-frontend/src/components/TodoList/TodoList.js":"3","/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItem/TodoItem.js":"4","/home/apoorv/todo-frontend/src/components/SignUp.js":"5","/home/apoorv/todo-frontend/src/components/TodoList/Footer/Footer.js":"6","/home/apoorv/todo-frontend/src/index.js":"7","/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItems.js":"8","/home/apoorv/todo-frontend/src/components/TodoList/AddTodo/AddTodo.js":"9","/home/apoorv/todo-frontend/src/components/Header.js":"10","/home/apoorv/todo-frontend/src/Api.js":"11"},{"size":1230,"mtime":1610119608941,"results":"12","hashOfConfig":"13"},{"size":1953,"mtime":1610120260489,"results":"14","hashOfConfig":"13"},{"size":2456,"mtime":1610120170948,"results":"15","hashOfConfig":"13"},{"size":1097,"mtime":1610120209961,"results":"16","hashOfConfig":"13"},{"size":2340,"mtime":1610119994849,"results":"17","hashOfConfig":"13"},{"size":1118,"mtime":1610120139064,"results":"18","hashOfConfig":"13"},{"size":254,"mtime":1610105084364,"results":"19","hashOfConfig":"13"},{"size":491,"mtime":1610120187400,"results":"20","hashOfConfig":"13"},{"size":927,"mtime":1610120424595,"results":"21","hashOfConfig":"13"},{"size":455,"mtime":1610119844047,"results":"22","hashOfConfig":"13"},{"size":2436,"mtime":1610119453585,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1mr5qeb",{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/apoorv/todo-frontend/src/App.js",[],"/home/apoorv/todo-frontend/src/components/Home.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/TodoList.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItem/TodoItem.js",[],"/home/apoorv/todo-frontend/src/components/SignUp.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/Footer/Footer.js",[],"/home/apoorv/todo-frontend/src/index.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItems.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/AddTodo/AddTodo.js",[],"/home/apoorv/todo-frontend/src/components/Header.js",[],"/home/apoorv/todo-frontend/src/Api.js",[]]
+[{"/home/apoorv/todo-frontend/src/App.js":"1","/home/apoorv/todo-frontend/src/components/Home.js":"2","/home/apoorv/todo-frontend/src/components/TodoList/TodoList.js":"3","/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItem/TodoItem.js":"4","/home/apoorv/todo-frontend/src/components/SignUp.js":"5","/home/apoorv/todo-frontend/src/components/TodoList/Footer/Footer.js":"6","/home/apoorv/todo-frontend/src/index.js":"7","/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItems.js":"8","/home/apoorv/todo-frontend/src/components/TodoList/AddTodo/AddTodo.js":"9","/home/apoorv/todo-frontend/src/components/Header.js":"10","/home/apoorv/todo-frontend/src/Api.js":"11","/home/apoorv/todo-frontend/src/components/Login.js":"12"},{"size":1230,"mtime":1610119608941,"results":"13","hashOfConfig":"14"},{"size":659,"mtime":1610121988645,"results":"15","hashOfConfig":"14"},{"size":2456,"mtime":1610120570941,"results":"16","hashOfConfig":"14"},{"size":1097,"mtime":1610120209961,"results":"17","hashOfConfig":"14"},{"size":2340,"mtime":1610119994849,"results":"18","hashOfConfig":"14"},{"size":1118,"mtime":1610120139064,"results":"19","hashOfConfig":"14"},{"size":254,"mtime":1610105084364,"results":"20","hashOfConfig":"14"},{"size":491,"mtime":1610120187400,"results":"21","hashOfConfig":"14"},{"size":927,"mtime":1610120424595,"results":"22","hashOfConfig":"14"},{"size":455,"mtime":1610119844047,"results":"23","hashOfConfig":"14"},{"size":2436,"mtime":1610119453585,"results":"24","hashOfConfig":"14"},{"size":1505,"mtime":1610121694098,"results":"25","hashOfConfig":"14"},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1mr5qeb",{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"36","messages":"37","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"38","messages":"39","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"40","messages":"41","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"44","messages":"45","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"48","messages":"49","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/apoorv/todo-frontend/src/App.js",[],"/home/apoorv/todo-frontend/src/components/Home.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/TodoList.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItem/TodoItem.js",[],"/home/apoorv/todo-frontend/src/components/SignUp.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/Footer/Footer.js",[],"/home/apoorv/todo-frontend/src/index.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/TodoItems/TodoItems.js",[],"/home/apoorv/todo-frontend/src/components/TodoList/AddTodo/AddTodo.js",[],"/home/apoorv/todo-frontend/src/components/Header.js",[],"/home/apoorv/todo-frontend/src/Api.js",[],"/home/apoorv/todo-frontend/src/components/Login.js",[]]
